finishing work project

+++ common.h
@@ -29,4 +29,4 @@ int address_parse(const char *address, const char *port, struct sockaddr_storage
 
 double degrees_to_radians(double degrees);
 double haversine(double lat1, double lon1, double lat2, double lon2);
-double distance_between_coordinates(Coordinate coord1, Coordinate coord2);
+int distance_between_coordinates(Coordinate coord1, Coordinate coord2);
